+++ grails-app/conf/application.groovy
@@ -2,9 +2,9 @@
 grails.plugin.springsecurity.logout.postOnly = false //REMOVER
 
 // Added by the Spring Security Core plugin:
-grails.plugin.springsecurity.userLookup.userDomainClassName = 'com.mini.asaas.User'
-grails.plugin.springsecurity.userLookup.authorityJoinClassName = 'com.mini.asaas.UserRole'
-grails.plugin.springsecurity.authority.className = 'com.mini.asaas.Role'
+grails.plugin.springsecurity.userLookup.userDomainClassName = 'com.mini.asaas.user.User'
+grails.plugin.springsecurity.userLookup.authorityJoinClassName = 'com.mini.asaas.user.UserRole'
+grails.plugin.springsecurity.authority.className = 'com.mini.asaas.user.Role'
 
 grails.plugin.springsecurity.controllerAnnotations.staticRules = [
 	[pattern: '/',               access: ['permitAll']],
