+++ grails-app/conf/spring/resources.groovy
@@ -1,4 +1,4 @@
-import com.mini.asaas.UserPasswordEncoderListener
+import com.mini.asaas.user.UserPasswordEncoderListener
 // Place your Spring DSL code here
 beans = {
     userPasswordEncoderListener(UserPasswordEncoderListener)
