--- grails-app/controllers/com/mini/asaas/SecureController.groovy
@@ -1,10 +1,0 @@
-package com.mini.asaas
-
-import grails.plugin.springsecurity.annotation.Secured
-
-class SecureController {
-   @Secured('ROLE_ADMIN')
-   def index() {
-      render 'Secure access only'
-   }
-}
