+++ grails-app/domain/com/mini/asaas/user/User.groovy
@@ -1,13 +1,15 @@
-package com.mini.asaas
+package com.mini.asaas.user
 
 import groovy.transform.EqualsAndHashCode
 import groovy.transform.ToString
 import grails.compiler.GrailsCompileStatic
+import com.mini.asaas.customer.Customer
+import com.mini.asaas.base.BaseEntity
 
 @GrailsCompileStatic
 @EqualsAndHashCode(includes='username')
 @ToString(includes='username', includeNames=true, includePackage=false)
-class User implements Serializable {
+class User extends BaseEntity implements Serializable {
 
     private static final long serialVersionUID = 1
 
@@ -18,6 +20,8 @@ class User implements Serializable {
     boolean accountLocked
     boolean passwordExpired
 
+    Customer customer
+
     Set<Role> getAuthorities() {
         (UserRole.findAllByUser(this) as List<UserRole>)*.role as Set<Role>
     }
@@ -25,6 +29,7 @@ class User implements Serializable {
     static constraints = {
         password nullable: false, blank: false, password: true
         username nullable: false, blank: false, unique: true
+        customer nullable: true, blank: true
     }
 
     static mapping = {
