+++ src/main/groovy/com/mini/asaas/UserPasswordEncoderListener.groovy
@@ -1,4 +1,4 @@
-package com.mini.asaas
+package com.mini.asaas.user
 
 import grails.plugin.springsecurity.SpringSecurityService
 import org.grails.datastore.mapping.engine.event.AbstractPersistenceEvent
